% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_unit_value2.R
\name{clean_value_unit_string}
\alias{clean_value_unit_string}
\title{Clean strings before value/unit parsing}
\usage{
clean_value_unit_string(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector with common formatting problems normalized.
}
\description{
Clean strings before value/unit parsing
}
