% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{conversions_mass}
\alias{conversions_mass}
\title{A conversion table for mass units}
\format{
\subsection{\code{conversions_mass}}{

A data frame with 6 rows and 3 columns:
\describe{
\item{from}{Units to transform from}
\item{to}{Units to transform to}
\item{factor}{number to multiply the value to tranform from one unit to the other}
}
}
}
\usage{
conversions_mass
}
\description{
A conversion table for mass units
}
\keyword{datasets}
