% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_unit_value2.R
\name{detect_context}
\alias{detect_context}
\title{Detect likely context/commentary mixed into the unit string}
\usage{
detect_context(u)
}
\arguments{
\item{u}{Character vector.}
}
\value{
Logical vector.
}
\description{
Detect likely context/commentary mixed into the unit string
}
