% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_unit_value2.R
\name{detect_non_physical}
\alias{detect_non_physical}
\title{Detect likely non-physical indicators / indices}
\usage{
detect_non_physical(u)
}
\arguments{
\item{u}{Character vector.}
}
\value{
Logical vector.
}
\description{
Detect likely non-physical indicators / indices
}
