% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_unit_value2.R
\name{detect_range}
\alias{detect_range}
\title{Detect entries that look like numeric ranges}
\usage{
detect_range(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Logical vector.
}
\description{
Detect entries that look like numeric ranges
}
