% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_unit_value2.R
\name{detect_suspicious_unit}
\alias{detect_suspicious_unit}
\title{Detect suspicious units that likely need manual review}
\usage{
detect_suspicious_unit(u)
}
\arguments{
\item{u}{Character vector.}
}
\value{
Logical vector.
}
\description{
Detect suspicious units that likely need manual review
}
