% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_units.R
\name{extract_denominator}
\alias{extract_denominator}
\title{Extract canonical denominator fragment}
\usage{
extract_denominator(x)
}
\arguments{
\item{x}{Character vector of units.}
}
\value{
Character vector such as \code{""}, \code{"/yr"}, \code{"/cap"}, \code{"/cap/yr"},
\code{"/ha"}, \code{"/ha/yr"}, or \code{"/km2"}.
}
\description{
Extract canonical denominator fragment
}
\keyword{internal}
