% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_unit_value.R
\name{extract_unit}
\alias{extract_unit}
\title{Extracts units from a string with values and Units}
\usage{
extract_unit(x)
}
\arguments{
\item{x}{a dcharacter vector that contains a value and a unit}
}
\value{
returns a vector with the units of the string as a character vector
}
\description{
The extract_unit function, takes a vector that contain units and values,
extracts only the value
}
\examples{
data("Nitrogen_Data")
extract_unit(Nitrogen_Data$boundary_percapita)
}
\author{
Derek Corcoran \href{mailto:derek.corcoran.barrios@gmail.com}{derek.corcoran.barrios@gmail.com}

Giorgia Graells \href{mailto:gygraell@gmail.com}{gygraell@gmail.com}
}
