% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_unit_value2.R
\name{extract_unit2}
\alias{extract_unit2}
\title{Extract unit text from strings with value + unit}
\usage{
extract_unit2(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector with extracted unit text.
}
\description{
Extract unit text from strings with value + unit
}
