% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_unit_value.R
\name{extract_value}
\alias{extract_value}
\title{Extracts values from a string with values and Units}
\usage{
extract_value(x)
}
\arguments{
\item{x}{a character vector that contains a value and a unit}
}
\value{
returns a vector with the values of the string as a numeric vector
}
\description{
The extract_value function, takes a vector that contain units and values,
extracts only the value
}
\examples{
data("Nitrogen_Data")
extract_value(Nitrogen_Data$boundary_percapita)
}
\author{
Derek Corcoran \href{mailto:derek.corcoran.barrios@gmail.com}{derek.corcoran.barrios@gmail.com}

Giorgia Graells \href{mailto:gygraell@gmail.com}{gygraell@gmail.com}
}
