% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_unit_value2.R
\name{extract_value2}
\alias{extract_value2}
\title{Extract numeric values from strings with value + unit}
\usage{
extract_value2(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Numeric vector. Returns \code{NA} for ranges and non-parseable entries.
}
\description{
Extract numeric values from strings with value + unit
}
