% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_unit_value2.R
\name{harmonize_unit_text}
\alias{harmonize_unit_text}
\title{Harmonize unit text conservatively}
\usage{
harmonize_unit_text(u)
}
\arguments{
\item{u}{Character vector of extracted unit text.}
}
\value{
Character vector of harmonized units.
}
\description{
Harmonize unit text conservatively
}
