% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Nitrogen_Data}
\alias{Nitrogen_Data}
\title{A dataset for the Nitrogen planetary bounderies}
\format{
\subsection{\code{Nitrogen_Data}}{

A data frame with 576 rows and 23 columns:
\describe{
\item{authors}{Authors of the publication}
\item{title}{Title of the publication}
\item{year}{Year}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
Nitrogen_Data
}
\description{
A subset of data from the review of planetary boundares
}
\keyword{datasets}
