% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_units.R
\name{normalize_substance_token}
\alias{normalize_substance_token}
\title{Normalize parsed substance token}
\usage{
normalize_substance_token(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector.
}
\description{
Normalize parsed substance token
}
\keyword{internal}
