% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_units.R
\name{parse_canonical_unit}
\alias{parse_canonical_unit}
\title{Parse harmonizable units into canonical components}
\usage{
parse_canonical_unit(u)
}
\arguments{
\item{u}{Character vector of already text-harmonized units.}
}
\value{
A data frame with columns:
\itemize{
\item \code{family}
\item \code{substance}
\item \code{denominator}
\item \code{multiplier}
}
}
\description{
Parse harmonizable units into canonical components
}
\keyword{internal}
