% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_unit_value2.R
\name{same_text}
\alias{same_text}
\title{Compare two character vectors after trimming and NA handling}
\usage{
same_text(x, y)
}
\arguments{
\item{x}{Character vector.}

\item{y}{Character vector.}
}
\value{
Logical vector.
}
\description{
Compare two character vectors after trimming and NA handling
}
\keyword{internal}
