% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_unit_value.R
\name{separate_unit_value}
\alias{separate_unit_value}
\title{Separates units from values in a dataframe}
\usage{
separate_unit_value(df, columns)
}
\arguments{
\item{df}{a dataframe which contains the columns to separate}

\item{columns}{a character with the options Mostconnected, Leastconnected and Ordered}
}
\value{
returns the same dataframe with the new columns
}
\description{
The separate_unit_value function, takes a column or columns from a dataframe that contain units and value,
. and separates it into two separate columns names \code{column_name_value} and \code{column_name_unit}
}
\examples{
# Mostconnected example
data("Nitrogen_Data")
separate_unit_value(df = Nitrogen_Data, columns = "boundary_percapita")
}
\author{
Derek Corcoran \href{mailto:derek.corcoran.barrios@gmail.com}{derek.corcoran.barrios@gmail.com}

Giorgia Graells \href{mailto:gygraell@gmail.com}{gygraell@gmail.com}
}
