% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_units.R
\name{strip_denominator}
\alias{strip_denominator}
\title{Remove canonical denominator fragment from unit text}
\usage{
strip_denominator(x)
}
\arguments{
\item{x}{Character vector of units.}
}
\value{
Character vector of unit cores.
}
\description{
Remove canonical denominator fragment from unit text
}
\keyword{internal}
