% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_unit_value.R
\name{separate_unit_value}
\alias{separate_unit_value}
\title{Separates units from values in a dataframe}
\usage{
separate_unit_value(
  df,
  columns = c("territorial_footprint", "territorial_boundary", "footprint_percapita",
    "boundary_percapita")
)
}
\arguments{
\item{df}{A dataframe which contains the columns to separate}

\item{columns}{A character vector with the names of columns to separate, it
defaults to}
}
\value{
Returns the same dataframe with the new columns appended
}
\description{
The separate_unit_value function takes one or more columns from a dataframe that
contain a value and a unit, and creates two new columns per input column:
\verb{\{column\}_value} and \verb{\{column\}_unit}, without removing the original column.
}
\examples{
data("Nitrogen_Data")
# By default it will separate columns  territorial_footprint, territorial_boundary,
# footprint_percapita, boundary_percapita
separate_unit_value(df = Nitrogen_Data)
# but can be changed if needed with more than one variable
separate_unit_value(df = Nitrogen_Data,
                columns = c("global_limit_considered","boundary_percapita"))
}
\author{
Derek Corcoran \href{mailto:derek.corcoran.barrios@gmail.com}{derek.corcoran.barrios@gmail.com}

Giorgia Graells \href{mailto:gygraell@gmail.com}{gygraell@gmail.com}
}
